+++ packages/std/src/math/decimal.rs
@@ -271,14 +271,14 @@ impl Decimal {
         self.0
             .checked_add(other.0)
             .map(Self)
-            .map_err(|_| OverflowError::new(OverflowOperation::Add, self, other))
+            .map_err(|_| OverflowError::new(OverflowOperation::Add))
     }
 
     pub fn checked_sub(self, other: Self) -> Result<Self, OverflowError> {
         self.0
             .checked_sub(other.0)
             .map(Self)
-            .map_err(|_| OverflowError::new(OverflowOperation::Sub, self, other))
+            .map_err(|_| OverflowError::new(OverflowOperation::Sub))
     }
 
     /// Multiplies one `Decimal` by another, returning an `OverflowError` if an overflow occurred.
@@ -288,11 +288,7 @@ impl Decimal {
         result_as_uint256
             .try_into()
             .map(Self)
-            .map_err(|_| OverflowError {
-                operation: OverflowOperation::Mul,
-                operand1: self.to_string(),
-                operand2: other.to_string(),
-            })
+            .map_err(|_| OverflowError::new(OverflowOperation::Mul))
     }
 
     /// Raises a value to the power of `exp`, panics if an overflow occurred.
@@ -330,11 +326,7 @@ impl Decimal {
             Ok(x * y)
         }
 
-        inner(self, exp).map_err(|_| OverflowError {
-            operation: OverflowOperation::Pow,
-            operand1: self.to_string(),
-            operand2: exp.to_string(),
-        })
+        inner(self, exp).map_err(|_| OverflowError::new(OverflowOperation::Pow))
     }
 
     pub fn checked_div(self, other: Self) -> Result<Self, CheckedFromRatioError> {
@@ -345,7 +337,7 @@ impl Decimal {
         self.0
             .checked_rem(other.0)
             .map(Self)
-            .map_err(|_| DivideByZeroError::new(self))
+            .map_err(|_| DivideByZeroError)
     }
 
     /// Returns the approximate square root as a Decimal.
@@ -1512,11 +1504,7 @@ mod tests {
     fn decimal_checked_mul_overflow() {
         assert_eq!(
             Decimal::MAX.checked_mul(Decimal::percent(200)),
-            Err(OverflowError {
-                operation: OverflowOperation::Mul,
-                operand1: Decimal::MAX.to_string(),
-                operand2: Decimal::percent(200).to_string(),
-            })
+            Err(OverflowError::new(OverflowOperation::Mul))
         );
     }
 
@@ -1788,11 +1776,7 @@ mod tests {
     fn decimal_checked_pow_overflow() {
         assert_eq!(
             Decimal::MAX.checked_pow(2),
-            Err(OverflowError {
-                operation: OverflowOperation::Pow,
-                operand1: Decimal::MAX.to_string(),
-                operand2: "2".to_string(),
-            })
+            Err(OverflowError::new(OverflowOperation::Pow))
         );
     }