+++ packages/std/src/math/decimal256.rs
@@ -280,14 +280,14 @@ impl Decimal256 {
         self.0
             .checked_add(other.0)
             .map(Self)
-            .map_err(|_| OverflowError::new(OverflowOperation::Add, self, other))
+            .map_err(|_| OverflowError::new(OverflowOperation::Add))
     }
 
     pub fn checked_sub(self, other: Self) -> Result<Self, OverflowError> {
         self.0
             .checked_sub(other.0)
             .map(Self)
-            .map_err(|_| OverflowError::new(OverflowOperation::Sub, self, other))
+            .map_err(|_| OverflowError::new(OverflowOperation::Sub))
     }
 
     /// Multiplies one `Decimal256` by another, returning an `OverflowError` if an overflow occurred.
@@ -297,11 +297,7 @@ impl Decimal256 {
         result_as_uint512
             .try_into()
             .map(Self)
-            .map_err(|_| OverflowError {
-                operation: OverflowOperation::Mul,
-                operand1: self.to_string(),
-                operand2: other.to_string(),
-            })
+            .map_err(|_| OverflowError::new(OverflowOperation::Mul))
     }
 
     /// Raises a value to the power of `exp`, panics if an overflow occurred.
@@ -339,11 +335,7 @@ impl Decimal256 {
             Ok(x * y)
         }
 
-        inner(self, exp).map_err(|_| OverflowError {
-            operation: OverflowOperation::Pow,
-            operand1: self.to_string(),
-            operand2: exp.to_string(),
-        })
+        inner(self, exp).map_err(|_| OverflowError::new(OverflowOperation::Pow))
     }
 
     pub fn checked_div(self, other: Self) -> Result<Self, CheckedFromRatioError> {
@@ -354,7 +346,7 @@ impl Decimal256 {
         self.0
             .checked_rem(other.0)
             .map(Self)
-            .map_err(|_| DivideByZeroError::new(self))
+            .map_err(|_| DivideByZeroError)
     }
 
     /// Returns the approximate square root as a Decimal256.
@@ -1574,11 +1566,7 @@ mod tests {
     fn decimal256_checked_mul_overflow() {
         assert_eq!(
             Decimal256::MAX.checked_mul(Decimal256::percent(200)),
-            Err(OverflowError {
-                operation: OverflowOperation::Mul,
-                operand1: Decimal256::MAX.to_string(),
-                operand2: Decimal256::percent(200).to_string(),
-            })
+            Err(OverflowError::new(OverflowOperation::Mul))
         );
     }
 
@@ -1863,11 +1851,7 @@ mod tests {
     fn decimal256_checked_pow_overflow() {
         assert_eq!(
             Decimal256::MAX.checked_pow(2),
-            Err(OverflowError {
-                operation: OverflowOperation::Pow,
-                operand1: Decimal256::MAX.to_string(),
-                operand2: "2".to_string(),
-            })
+            Err(OverflowError::new(OverflowOperation::Pow))
         );
     }