+++ packages/std/src/math/uint256.rs
@@ -233,35 +233,35 @@ impl Uint256 {
         self.0
             .checked_add(other.0)
             .map(Self)
-            .ok_or_else(|| OverflowError::new(OverflowOperation::Add, self, other))
+            .ok_or_else(|| OverflowError::new(OverflowOperation::Add))
     }
 
     pub fn checked_sub(self, other: Self) -> Result<Self, OverflowError> {
         self.0
             .checked_sub(other.0)
             .map(Self)
-            .ok_or_else(|| OverflowError::new(OverflowOperation::Sub, self, other))
+            .ok_or_else(|| OverflowError::new(OverflowOperation::Sub))
     }
 
     pub fn checked_mul(self, other: Self) -> Result<Self, OverflowError> {
         self.0
             .checked_mul(other.0)
             .map(Self)
-            .ok_or_else(|| OverflowError::new(OverflowOperation::Mul, self, other))
+            .ok_or_else(|| OverflowError::new(OverflowOperation::Mul))
     }
 
     pub fn checked_pow(self, exp: u32) -> Result<Self, OverflowError> {
         self.0
             .checked_pow(exp)
             .map(Self)
-            .ok_or_else(|| OverflowError::new(OverflowOperation::Pow, self, exp))
+            .ok_or_else(|| OverflowError::new(OverflowOperation::Pow))
     }
 
     pub fn checked_div(self, other: Self) -> Result<Self, DivideByZeroError> {
         self.0
             .checked_div(other.0)
             .map(Self)
-            .ok_or_else(|| DivideByZeroError::new(self))
+            .ok_or(DivideByZeroError)
     }
 
     pub fn checked_div_euclid(self, other: Self) -> Result<Self, DivideByZeroError> {
@@ -272,12 +272,12 @@ impl Uint256 {
         self.0
             .checked_rem(other.0)
             .map(Self)
-            .ok_or_else(|| DivideByZeroError::new(self))
+            .ok_or(DivideByZeroError)
     }
 
     pub fn checked_shr(self, other: u32) -> Result<Self, OverflowError> {
         if other >= 256 {
-            return Err(OverflowError::new(OverflowOperation::Shr, self, other));
+            return Err(OverflowError::new(OverflowOperation::Shr));
         }
 
         Ok(Self(self.0.shr(other)))
@@ -285,7 +285,7 @@ impl Uint256 {
 
     pub fn checked_shl(self, other: u32) -> Result<Self, OverflowError> {
         if other >= 256 {
-            return Err(OverflowError::new(OverflowOperation::Shl, self, other));
+            return Err(OverflowError::new(OverflowOperation::Shl));
         }
 
         Ok(Self(self.0.shl(other)))
@@ -1111,11 +1111,7 @@ mod tests {
         let target = Uint128::try_from(source);
         assert_eq!(
             target,
-            Err(ConversionOverflowError::new(
-                "Uint256",
-                "Uint128",
-                Uint256::MAX.to_string()
-            ))
+            Err(ConversionOverflowError::new("Uint256", "Uint128"))
         );
     }
 
@@ -1354,10 +1350,8 @@ mod tests {
 
         // error result on underflow (- would produce negative result)
         let underflow_result = a.checked_sub(b);
-        let OverflowError {
-            operand1, operand2, ..
-        } = underflow_result.unwrap_err();
-        assert_eq!((operand1, operand2), (a.to_string(), b.to_string()));
+        let OverflowError { operation } = underflow_result.unwrap_err();
+        assert_eq!(operation, OverflowOperation::Sub);
     }
 
     #[test]
@@ -1814,9 +1808,6 @@ mod tests {
             Err(ConversionOverflow(ConversionOverflowError {
                 source_type: "Uint512",
                 target_type: "Uint256",
-                value:
-                    "303954234247955012986873835647805758114833709747306480603576158020771965304829"
-                        .to_string()
             })),
         );
     }
@@ -1833,9 +1824,7 @@ mod tests {
         let fraction = (21u128, 0u128);
         assert_eq!(
             Uint256::from(123456u32).checked_mul_floor(fraction),
-            Err(DivideByZero(DivideByZeroError {
-                operand: "2592576".to_string()
-            })),
+            Err(DivideByZero(DivideByZeroError)),
         );
     }
 
@@ -1909,9 +1898,6 @@ mod tests {
             Err(ConversionOverflow(ConversionOverflowError {
                 source_type: "Uint512",
                 target_type: "Uint256",
-                value:
-                    "303954234247955012986873835647805758114833709747306480603576158020771965304829" // raises prior to rounding up
-                        .to_string()
             })),
         );
     }
@@ -1928,9 +1914,7 @@ mod tests {
         let fraction = (21u128, 0u128);
         assert_eq!(
             Uint256::from(123456u32).checked_mul_ceil(fraction),
-            Err(DivideByZero(DivideByZeroError {
-                operand: "2592576".to_string()
-            })),
+            Err(DivideByZero(DivideByZeroError)),
         );
     }
 
@@ -2003,9 +1987,6 @@ mod tests {
             Err(ConversionOverflow(ConversionOverflowError {
                 source_type: "Uint512",
                 target_type: "Uint256",
-                value:
-                    "303954234247955012986873835647805758114833709747306480603576158020771965304829"
-                        .to_string()
             })),
         );
     }
@@ -2079,9 +2060,6 @@ mod tests {
             Err(ConversionOverflow(ConversionOverflowError {
                 source_type: "Uint512",
                 target_type: "Uint256",
-                value:
-                    "303954234247955012986873835647805758114833709747306480603576158020771965304829"
-                        .to_string() // raises prior to rounding up
             })),
         );
     }