+++ packages/storage/src/singleton.rs
@@ -262,13 +262,8 @@ mod tests {
         };
         writer.save(&cfg).unwrap();
 
-        let output = writer.update(|_c| {
-            Err(StdError::from(OverflowError::new(
-                OverflowOperation::Sub,
-                4,
-                7,
-            )))
-        });
+        let output =
+            writer.update(|_c| Err(StdError::from(OverflowError::new(OverflowOperation::Sub))));
         match output.unwrap_err() {
             StdError::Overflow { .. } => {}
             err => panic!("Unexpected error: {err:?}"),